refactor: we no longer rely on the Replicator for the currently selected Actor
notes: instead, we use the existing 'PreviouslySelectedActor' variable to select the Actor. This avoids calling the Replicator which is a soft-lock-crash if called from a Client. Since we _want_ to debug Actors on Clients, this change makes it possible without making an Engine modification. The downside is that we lose the Actor name in the top bar. However, that is easily remedied by providing the name in the gameplay debug text as part of our Actor Submenu